Renamed UBO to VulkanUBO, and created UBO base class to be used in both Vulkan and OpenGL
