+++ src/main/kotlin/graphics/scenery/backends/opengl/OpenGLObjectState.kt
@@ -3,6 +3,7 @@ package graphics.scenery.backends.opengl
 import cleargl.GLProgram
 import cleargl.GLTexture
 import graphics.scenery.NodeMetadata
+import java.nio.ByteBuffer
 import java.util.*
 import java.util.concurrent.ConcurrentHashMap
 
@@ -14,6 +15,7 @@ import java.util.concurrent.ConcurrentHashMap
  *  default consumers.
  */
 class OpenGLObjectState : NodeMetadata {
+    data class OpenGLBufferBinding(var buffer: ByteBuffer? = null, var offset: Long = 0L, var name: String, var binding: Int = 0)
     /** List of consumers of this metadata, e.g. [OpenGLRenderer] */
     override val consumers: MutableList<String> = ArrayList<String>()
 
@@ -40,6 +42,8 @@ class OpenGLObjectState : NodeMetadata {
     var mStoredIndexCount = 0
     /** Number of stored vertex/normal/texcoord primitives. */
     var mStoredPrimitiveCount = 0
+    /** buffer bindings **/
+    var bufferBindings = ArrayList<OpenGLBufferBinding>()
 
     /**
      * Default constructor, adding the [OpenGLRenderer]
